DSEGOG-57 improving on date line annotations, sorting plot data by x-axis ascending
Improving on the function to extract all the dates present in the datasets. If we have a data point for April 27, the midnight timestamp for April 27 is returned

Also ensuring we sort the plot data query by x-axis ascending. This slows down the plot due to it needing to re-query whenever the x-axis changes but is the only way to ensure the line chart is correct when plotting channel vs channel